Cleaned up the project
Previously, the project was a bit of a mess. We've removed redundant files, refactored things slightly, and overall made changes that reflect a better structure.
