--- machine-player/DeepQNetwork.py
@@ -1,34 +1,0 @@
-# Imports
-import torch as T
-import torch.nn as nn
-import torch.nn.functional as F
-import torch.optim as optim
-import numpy as np
-
-# Define the DeepQNetwork class
-class DeepQNetwork(nn.Module):
-    def __init__(self):
-        super(DeepQNetwork, self).__init__()
-        lr = 0.001
-        input_dims = (525,)
-        fc1_dims = 256
-        fc2_dims = 256
-        n_actions = 382
-        self.input_dims = input_dims
-        self.fc1_dims = fc1_dims
-        self.fc2_dims = fc2_dims
-        self.n_actions = n_actions
-        self.fc1 = nn.Linear(*self.input_dims, self.fc1_dims)
-        self.fc2 = nn.Linear(self.fc1_dims, self.fc2_dims)
-        self.fc3 = nn.Linear(self.fc2_dims, self.n_actions)
-        self.optimizer = optim.Adam(self.parameters(), lr=lr)
-        self.loss = nn.MSELoss()
-        self.device = T.device("cuda:0" if T.cuda.is_available() else "cpu")
-        self.to(self.device)
-
-    # Define the forward pass through the network
-    def forward(self, state):
-        x = F.relu(self.fc1(state))
-        x = F.relu(self.fc2(x))
-        actions = self.fc3(x)
-        return actions
