+++ machine-player/TrainingSession.py
@@ -8,9 +8,7 @@ from datetime import datetime
 from CatanGame import CatanGame
 from Randy import Randy
 from Adam import Adam
-from Phil import Phil
 from Redmond import Redmond
-from Eugene import Eugene
 
 # Define the TrainingSession class
 class TrainingSession:
@@ -122,14 +120,7 @@ class TrainingSession:
                 memory_tuple = (game_state, action_index, reward, new_game_state, done)
 
                 # Feed the memory tuple to the agent
-                if self.AGENT_SELECTED == "Phil":
-                    # Find the index of the action in the list of all actions
-                    action_with_idx = all_actions.index(chosen_action)
-                    self.AGENT.store_transition(
-                        game_state, action_with_idx, reward, new_game_state, game_over
-                    )
-                else:
-                    self.AGENT.feed_memory(memory_tuple)
+                self.AGENT.feed_memory(memory_tuple)
 
                 # Increment the steps until call learning
                 self.steps_until_call_learning += 1
@@ -282,15 +273,6 @@ class TrainingSession:
                 print("So, you've awoken me again...\nRedmond reporting for duty!")
                 self.AGENT.load_model(self.MODEL_PATH)
                 agent_already_exists = True
-        elif self.AGENT_SELECTED == "Phil":
-            self.AGENT = Phil(
-                gamma=0.9,
-                epsilon=1.0,
-                lr=0.001,
-                input_dims=(525,),
-                batch_size=64,
-                n_action=382,
-            )
         else:
             self.AGENT = Randy()
