--- machine-player/main.py
@@ -1,18 +1,0 @@
-#!/usr/bin/env python3
-"""
-Module Docstring
-"""
-
-__author__ = "Reef Lakin"
-__version__ = "0.1.0"
-__license__ = "MIT"
-
-
-def main():
-    """ Main entry point of the app """
-    print("hello world")
-
-
-if __name__ == "__main__":
-    """ This is executed when run from the command line """
-    main()
